TensorExpr eval: fix copying variables on big endian systems
Since there is a separate buffer present now, ensure it's properly filled.

This change fixes TestTensorExprPyBind::test_dynamic_shape
and TestTensorExprPyBind::test_dynamic_shape_2d tests
from test/test_tensorexpr_pybind.py on big endian systems.

+++ torch/csrc/jit/tensorexpr/tensorexpr_init.cpp
@@ -830,7 +830,7 @@ void initTensorExprBindings(PyObject* module) {
             value_ptrs.reserve(py::len(values));
             for (const auto& value : values) {
               if (py::isinstance<py::int_>(value)) {
-                value_ptrs.emplace_back(value.cast<int64_t>());
+                value_ptrs.emplace_back(value.cast<int>());
               } else {
                 value_ptrs.emplace_back(value.cast<at::Tensor>().data_ptr());
               }
